Add paddle collision
Currently, the ball is travelling through the paddle. Add a collision by checking wether the x position of the ball is between the start/end of the paddle and the y position is greater than the y position of the paddle. If this is true, then negate the y velocity of the ball.