Start game on spacebar
Presently, there is no active state in the game - when are we playing vs paused/game ended. We want to control the start of the game with the spacebar - the ball should start moving. To do this, add a boolean called playing to the state. In the update function, if the spacebar is pressed, check whether playing is set to false. If so, assign the ballPosition and ballVelocity.