Change ball to start in the center
The ball starts and resets around the top left rather than the center. This position is also hard coded. Add a startPosition variable to ballAttributes constant which can replace the hard coded values. Also change the x and y of the ball to be in the center of the game screen.