Test ball/brick collisions
Create a function 'doesBallHitBrick' which returns true if the filtered bricks and the current bricks in the model are not equal. This means that the ball is colliding with a brick but the model has not been updated. Add this function to updateBallVelocity and switch calling order in gameLoop so that bricks are updated after velocity is calculated.