Fix paddle ball collision
The bal/paddle collision detection is not accurate. The ball bounces slightly underneath the top of the paddle and makes contact unevenly on the paddle sides. To fix this, update the collision logic to use the center of the ball for the x direction and to include the height of the ball in the y direction.