Style bricks
The bricks are currently not visibile. Firstly change their positions and sizing to fit the game. TheChange their colours to reflect the number of points they are worth. This will require adding 'rank' to the brick model which, at initialization, attributes each brick a different rank according to it's position and then is styled (coloured) appropriately.