Fix gameState bugs
A few bugs exist from the introduction of gameState. 1. After unpausing the game, the ball always travels in the same direction. Store the velocity of the ball on pause. 2. The paddle is able to move during pause. 3. The score does not reset when a new game begins - having lost all lives.