Make paddle movement smooth
The paddle animation is not smooth currently. This is due to it being independent to the tick which uses the AnimationFrame library. To solve this, on key press add a 'velocity' to the paddle and on key up remove it. Then update paddle position within the tick loop using this velocity.